Module cargo_crev_web::html_server_template_mod

html_server_template_mod html templating library for the web server should be compatible also with svg, because of namespaces

Structs

Attribute
An attribute on a DOM node, such as id="my-thing"

ElementNode

SubTemplate

Enums

HtmlOrSvg
Svg elements are different because they have a namespace

Node

Traits

HtmlServerTemplateRender

Functions

decode_5_xml_control_characters
private fn - decode 5 xml control characters : " ' & < > https://www.liquid-technologies.com/XML/EscapingData.aspx I will ignore all html entities, to keep things simple, because all others characters can be written as utf-8 characters. it is mandatory that text is valid utf-8. https://www.tutorialspoint.com/html5/html5_entities.htm TODO: find a faster method // The standard library replace() function makes allocation,

decode_html_script_node
in html the