Trait cargo_crev_web::html_server_template_mod::HtmlServerTemplateRender

pub trait HtmlServerTemplateRender {
    fn render_html_file(&self, templates_folder_name: &str) -> String;

    fn data_model_name(&self) -> String;

    fn replace_with_string(
        &self, 
        placeholder: &str, 
        subtemplate: &str, 
        pos_cursor: usize
    ) -> String;

    fn replace_with_url(
        &self, 
        placeholder: &str, 
        subtemplate: &str, 
        pos_cursor: usize
    ) -> UrlUtf8EncodedString;

    fn retain_next_node_or_attribute(&self, placeholder: &str) -> bool;

    fn replace_with_nodes(&self, placeholder: &str) -> Vec<Node>;

    fn render_sub_template(
        &self, 
        template_name: &str, 
        sub_templates: &Vec<SubTemplate>
    ) -> Vec<Node>;

    fn render_from_file(&self, template_file_name: &str) -> String { ... }

    fn render(&self, html_template_raw: &str) -> String { ... }

    fn render_template_raw_to_nodes(
        &self, 
        html_template_raw: &str, 
        html_or_svg_parent: HtmlOrSvg, 
        subtemplate: &str, 
        pos_cursor: usize
    ) -> Result<Vec<Node>, &'static str> { ... }

    fn fill_element_node(
        &self, 
        reader_for_microxml: &mut ReaderForMicroXml, 
        element: ElementNode, 
        html_or_svg_parent: HtmlOrSvg, 
        dom_path: &mut Vec<String>, 
        sub_templates: &Vec<SubTemplate>, 
        subtemplate: &str, 
        pos_cursor: usize, 
        retain_this_node: bool
    ) -> Option<Result<ElementNode, &'static str>> { ... }

    fn extract_children_sub_templates(template_raw: &str) -> Vec<SubTemplate> { ... }

    fn root_element_node_to_html_string(
        element_node: &ElementNode
    ) -> Result<String, String> { ... }
}

Required methods

fn render_html_file(&self, templates_folder_name: &str) -> String

The code for templating starts here. renders the complete html file. Not a sub-template/fragment.

fn data_model_name(&self) -> String

name of data model for debugging

fn replace_with_string(
&self,
placeholder: &str,
subtemplate: &str,
pos_cursor: usize
) -> String

returns a String to replace the next text-node or attribute value use macro s!() for a normal string

fn replace_with_url(
&self,
placeholder: &str,
subtemplate: &str,
pos_cursor: usize
) -> UrlUtf8EncodedString

same as replace_with_string, but return url exclusively for attributes value of href and src the url must be encoded in the beginning because it encodes segments of url prior to being composed together. use macro url_u!() to create an url, very like format! I try to avoid String here to force the developer to not forget to url_encode

fn retain_next_node_or_attribute(&self, placeholder: &str) -> bool

boolean : is the next node rendered or not

fn replace_with_nodes(&self, placeholder: &str) -> Vec<Node>

returns a vector of Nodes to replace the next Node

fn render_sub_template(
&self,
template_name: &str,
sub_templates: &Vec<SubTemplate>
) -> Vec<Node>

renders sub-template

Loading content...
Provided methods

fn render_from_file(&self, template_file_name: &str) -> String

render root template (not sub-templates) from file

fn render(&self, html_template_raw: &str) -> String

render for root template (not subtemplates) from string

fn render_template_raw_to_nodes(
&self,
html_template_raw: &str,
html_or_svg_parent: HtmlOrSvg,
subtemplate: &str,
pos_cursor: usize
) -> Result<Vec<Node>, &'static str>

this is used for templates and subtemplates equally first extracts all children sub_templates returns Nodes

fn fill_element_node(
&self,
reader_for_microxml: &mut ReaderForMicroXml,
element: ElementNode,
html_or_svg_parent: HtmlOrSvg,
dom_path: &mut Vec<String>,
sub_templates: &Vec<SubTemplate>,
subtemplate: &str,
pos_cursor: usize,
retain_this_node: bool
) -> Option<Result<ElementNode, &'static str>>

fn extract_children_sub_templates(template_raw: &str) -> Vec<SubTemplate>

extracts and saves sub_templates only one level deep: children

fn root_element_node_to_html_string(
element_node: &ElementNode
) -> Result<String, String>

converts element node to string the attribute values and Text nodes and Comments are xml encoded

Loading content...