Module cargo_crev_web::html_template_mod

html_template_mod

Structs

Attribute
An attribute on a DOM node, such as id="my-thing"

ElementNode

Node

SubTemplate

Enums

HtmlOrSvg
Svg elements are different because they have a namespace

NodeEnum

Traits

HtmlTemplating

Functions

element_node_to_string
default implementation - render template to string

template_raw_from_file
only the root template comes from a file all sub-templates later are raw strings

to_string_zero_to_empty