Struct cargo_crev_web::reserved_folder_mod::ReservedFolder

pub struct ReservedFolder {
    pub list_trusted_reviewer_id: Option<Vec<OnlyReviewer>>,
    pub reindex_after_fetch_new_reviews: Option<String>,
    pub fetch_new_reviews: Option<String>,
    pub blocklisted_repos: Option<Vec<(String, String)>>,
    pub list_new_reviewer_id: Option<Vec<OnlyReviewer>>,
    pub daily_visitors: Option<Vec<DailyVisitors>>,
}

Fields
list_trusted_reviewer_id: Option<Vec<OnlyReviewer>>reindex_after_fetch_new_reviews: Option<String>fetch_new_reviews: Option<String>blocklisted_repos: Option<Vec<(String, String)>>list_new_reviewer_id: Option<Vec<OnlyReviewer>>daily_visitors: Option<Vec<DailyVisitors>>
Implementations

impl ReservedFolder

pub fn new(_state_global: ArcMutStateGlobal) -> Self

prepares the data

pub fn list_trusted_reviewer_id(state_global: ArcMutStateGlobal) -> Self

pub fn reindex_after_fetch_new_reviews(state_global: ArcMutStateGlobal) -> Self

pub fn fetch_new_reviews(_state_global: ArcMutStateGlobal) -> Self

pub fn blocklisted_repos(_state_global: ArcMutStateGlobal) -> Self

fn fill_blocklisted_repos(&mut self)

read blocklisted_repos from json file

pub async fn list_new_reviewer_id(state_global: ArcMutStateGlobal) -> Self

The command cargo crev id query all returns a list of repos found in all trust crev files: mine (/home/me/.config/crev) and from all the remotes (/home/me/.cache/crev/remotes). I used for some time the repo https://gitlab.com/crev-dev/auto-crev-proofs.git. It finds new repos automatically, but it looks that it is not working any more. I will do it myself: find all forked crev-proofs on github and gitlab and cargo crev id query all. There are also some manually added repos in: https://github.com/crev-dev/cargo-crev/wiki/List-of-Proof-Repositories But it looks that this list is obsolete. I will add all the new repos with cargo crev trust --level low <url>. Ajooj: this automatically fetches the proofs with all transient. I don’t want that. Warning: cargo crev id trust --level low <hash> sounds very similar, but it accepts only hash, not url! And without url it is not possible to find the hash. For list only directly trusted identities: cargo crev id query trusted --high-cost 1 --medium-cost 1 --low-cost 1 --depth 1 This way I can then use cargo crev repo fetch trusted --high-cost 1 --medium-cost 1 --low-cost 1 --depth 1

Warning: never put passwords, passphrases ot tokens inside the code and publish it to Github ! If you wrote some secret in the bash like export PASS=my_secret, delete it from the history with this command: history | tac | grep export | cut -d ’ ’ -f 2 | awk ‘{printf “history -d “ $1 “; “}’

Warning: for cargo crev trust to work I need the crev passphrase. I will paste the passphrase from the clipboard to the env variable before starting the server: export CREV_PASSPHRASE=$(pbpaste) To use pbpaste and pbcopy in Debian on WSL2 use this instructions: https://www.techtronic.us/pbcopy-pbpaste-for-wsl/. But if clip.exe does not work you need to do this first to your /etc/profile file: https://github.com/microsoft/WSL/issues/5779#issuecomment-675574471 Use sudo nano /etc/profile. I rather comment the lines with #. And after that, restart wsl in Administrative mode Cmd in Win10: wsl -shutdown.

For unauthenticated requests on Github, the rate limit allows for up to 60 requests per hour. So I need to be authenticated for Github api with github PAT (personal access token) Before running the web server I store it in the environment variable:
export GITHUB_TOKEN=$(pbpaste)

pub fn daily_visitors(_state_global: ArcMutStateGlobal) -> Self

fn item_at_cursor_1(
&self,
subtemplate: &str,
pos_cursor: usize
) -> Option<&OnlyReviewer>

return the item at cursor or default

fn item_at_cursor_2(
&self,
subtemplate: &str,
pos_cursor: usize
) -> Option<&OnlyReviewer>

fn item_at_cursor_3(
&self,
subtemplate: &str,
pos_cursor: usize
) -> Option<&DailyVisitors>