Module cargo_crev_web::url_utf8_mod

url_utf8_mod url encoding and decoding on the web server This module is strictly limited to utf8 urls.
Url is made of parts, fragments or segments mostly delimited by slash “/”.
They must be separately encoded/decoded, not as a whole string.
It is impossible to guarantee that the whole string is correctly encoded/decoded.
But is possible to minimize the misuse of the String type for Url.
With the normal String it is not possible to force the developer to encode/decode.
With special wrapper types around String is possible to help the coder to write properly and not forget about it.
TODO: analyze if is possible to use more &str and Cow instead of always allocating String.
But urls are usually small and this is not a priority.

Structs

UrlPartUtf8Decoded

the url must be utf 8. Only the 5 control characters are encoded. url has parts or fragments or segments delimited mostly by slash / every part must be encoded/decoded separately, to maintain the control character slash /

UrlUtf8EncodedString

Type UrlUtf8EncodedString explicitly informs that the content has been url encoded. It contains a string with the whole url. The url is constructed with a special macro, where the dynamic parts are always encoded. It is impossible to force the developer to properly encode the static part of the url. But this special type is making this kind of errors difficult, obvious and traceable. TODO: the macro could use repetition to avoid having 4 fn with different number of parameters.

Constants

FRAGMENT 🔒

https://url.spec.whatwg.org/#fragment-percent-encode-set