Crate cargo_crev_web

main.rs Here will have only doc comments, because they are so very large. It is then difficult to code with all this doc comments cluttering the text editor. The actual code will go to a separate file/mod : main_code_mod.rs

Modules

author_reviews_mod
author_reviews_mod

authors_mod
authors_mod

badge_mod
badge_mod

crate_reviews_mod
crate_reviews_mod

crate_version_summary_mod
crate_version_summary_mod

crates_mod
crates_mod

data_file_scan_mod
data_file_scan_mod

encode_decode_mod
encode_decode_mod

html_server_template_mod
html_server_template_mod html templating library for the web server should be compatible also with svg, because of namespaces

issue_mod
issue_mod

macros_mod

main_code_mod
main.rs

reserved_folder_mod
reserved_folder_mod This is only one module/html page, but can execute different actions. The data model must have fields for every action as Option<>. Because only this data can influence the html render. There are different "new" functions for different actions, to prepare adequate data. If field is is_some(), then render the html part dedicated to this action.

review_index_mod
review_index_mod

review_index_summary_mod
review_index_summary_mod

review_mod
review_mod

review_new_mod
review_mod

router_mod
router_mod

utils_mod
utils_mod.rs

version_summary_mod
crate_version_summary_mod

Macros

s
short macro s! for String::from(&str). because that is so common. Equivalents: x.to_string(), x.to_owned(),...

Structs

Attribute
An attribute on a DOM node, such as id="my-thing"

ElementNode

SubTemplate

Enums

HtmlOrSvg
Svg elements are different because they have a namespace

Node

Traits

HtmlServerTemplateRender

Functions

author_name_from_url

conditional_usize
similar to ternary operator

find_from
find str from pos_cursor low level

find_pos_after_delimiter
return the position after the delimiter or None Does NOT mutate the pos_cursor, because that is for a higher level logic to decide.

find_pos_before_delimiter
return the position before the delimiter or None Does NOT mutate the pos_cursor, because that is for a higher level logic to decide.

find_range_between_delimiters
find and return the range of the first occurrence between start and end delimiters Success: mutates also the cursor position, so the next find will continue from there Fail: return None if not found and don't mutate pos_cursor I use type Range to avoid references &str and lifetimes. But the programmer can make the error to apply the range to the wrong vector.

main
main function of the binary

main_code

ns_elapsed

ns_print

ns_start

parse_semver
parse semver ex. 12.99.88alpha

render_sub_template_match_else
boilerplate

replace_with_nodes_match_else
boilerplate

replace_with_string_match_else
boilerplate

retain_next_node_or_attribute_match_else
boilerplate

to_string_zero_to_empty
to string, but zero converts to empty

traverse_dir_with_exclude_dir
traverse dir (sub-dir) with exclude dir the find_file and the exclude dir strings must start with /

Type Definitions

CachedReviewIndex